requirements.txt
streamlit
openai
llama-index
llama-index-llms-openai
nltk
pypdf
python-dotenv